Implement SEM realizations
Given a DAG and a gaussoid, print the system that is the intersection of
the realization space of the gaussoid with the structural equation model
defined by the DAG, in SEM coordinates. Currently only with independent
and normed noise (i.e. having only lambdas, no omegas).

For that I had to import my Graph::* packages. I should take the time
to bundle this into a nice installable, testable module very soon.