Scripts for looking at oriented realization spaces for SEMs
dag-model.sh gives for a given DAG each system that describes the SEM
intersected with the realization space of an orientation of the DAG's
gaussoid, ready for immediate bulk pasting into Mathematica.

dag-oriented.pl is a helper script that fetches orientations for a
gaussoid from an external file and makes fitting command lines for
grealize.

Also change the grealize formatting for Mathematica from "human-readable"
to "one line per system".