Enforce DAG vertex ordering
We die if the digraph has an edge that reveals a wrong ordering of the
vertices. Edges must only be $i -> $j where $i < $j so ensure that the
Lambda matrix is upper triangular.

Before this was not checked and the resulting system was be malformed,
because variables were omitted (when their place would be in the lower
triangle of Lambda).